refactor: zod type, scheme 정의
