+++ src/routes/order/-queryOptions.ts
@@ -1,25 +1,7 @@
 import { queryOptions } from '@tanstack/react-query';
 import { z } from 'zod';
 import fetcher from '../../lib/axios';
-import { Product, productSchema } from '../products/-queryOptions';
-
-interface OrderDetails extends Product {
-  quantity: number;
-}
-
-export interface Order {
-  id: number;
-  orderDetails: OrderDetails[];
-}
-
-const orderDetailSchema = productSchema.extend({
-  quantity: z.number(),
-});
-
-export const orderSchema = z.object({
-  id: z.number(),
-  orderDetails: z.array(orderDetailSchema),
-});
+import { Order, orderSchema } from '../../types';
 
 const fetchOrderList = async () => {
   const { data } = await fetcher.get<Order[]>('/orders');
